i3lock: lock screen before suspending

+++ arch/config/i3/i3lock_fancy.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
 # Pre lock (pause music and notifications)
-SPOTIFY_STATUS=$(playerctl -p spotify status)
-if [[ $SPOTIFY_STATUS == "Playing" ]]; then
-    playerctl -p spotify pause
+if [ "$1" != 'suspend' ]; then
+    SPOTIFY_STATUS=$(playerctl -p spotify status)
+    if [[ $SPOTIFY_STATUS == "Playing" ]]; then
+        playerctl -p spotify pause
+    fi
+    killall -SIGUSR1 dunst
 fi
-killall -SIGUSR1 dunst
 
 # Actual locking
 B='#24272EFF'  # inside color
@@ -43,4 +45,6 @@ i3lock \
     --radius=100
 
 # Post lock
-killall -SIGUSR2 dunst
+if [ "$1" != 'suspend' ]; then
+    killall -SIGUSR2 dunst
+fi
